Moved DateParserTest and geospatialUtilsJTS to common package.
